remove unused assignment: do not move the transformed statement up as it could change the semantic and is not evident (IDEA-123841)

+++ java/java-impl/src/com/intellij/codeInspection/defUse/DefUseInspection.java
@@ -118,8 +118,13 @@ public class DefUseInspection extends DefUseInspectionBase {
         else if (res == RemoveUnusedVariableUtil.MAKE_STATEMENT) {
           final PsiElementFactory factory = JavaPsiFacade.getInstance(project).getElementFactory();
           final PsiStatement statementFromText = factory.createStatementFromText(psiInitializer.getText() + ";", null);
-          declaration.getParent().addAfter(statementFromText, declaration);
-          elementToDelete.delete();
+          final PsiElement parent = elementToDelete.getParent();
+          if (parent instanceof PsiExpressionStatement) {
+            parent.replace(statementFromText);
+          } else {
+            declaration.getParent().addAfter(statementFromText, declaration);
+            elementToDelete.delete();
+          }
         }
       }
       catch (IncorrectOperationException e) {
